requirements.txt
streamlit
requests
google-cloud-discoveryengine
google-cloud-aiplatform==1.42.0
vertexai==1.42.0
functions-framework
rich<14
